requirements.txt
keras==2.9.0
numpy==1.17.4
Pillow==9.4.0
scikit_image==0.19.3
scikit_learn==1.2.0
streamlit==1.12.2
tensorflow==2.9.1
